requirements.txt
typer>=0.16,<1.0
rich>=13.7,<14.0
pydantic>=2.8,<3.0
pydantic-settings>=2.3,<3.0
PyYAML>=6.0,<7.0
gitpython>=3.1,<4.0
structlog>=24.1,<25.0
tenacity>=8.2,<9.0
questionary>=2.0,<3.0
langchain>=0.2,<0.3
langchain-community>=0.2,<0.3
langgraph>=0.0.59,<1.0
crewai>=0.28,<1.0
textual>=0.47,<1.0
